Revert "Merge branch 'main' into user-dto"
This reverts commit 8c3115bc134c523f86addb5dd97c4510f43c4da1, reversing
changes made to efc70d2639d6d4de19042f4e577281061f729d78.
