+++ backend/src/main/java/app/backend/controller/UserController.java
@@ -35,29 +35,29 @@ public class UserController {
 		return "Welcome this endpoint is not secure";
 	}
 
-	@PostMapping("/addNewUser")
+	@PostMapping("/signup")
 	public String addNewUser(@RequestBody UserInfo userInfo) {
 		return service.addUser(userInfo);
 	}
 
 	@GetMapping("/user/userProfile")
-	@PreAuthorize("hasAuthority('ROLE_USER')")
+	@PreAuthorize("hasAuthority('DEVELOPER')")
 	public String userProfile() {
 		return "Welcome to User Profile";
 	}
 
 	@GetMapping("/admin/adminProfile")
-	@PreAuthorize("hasAuthority('ROLE_ADMIN')")
+	@PreAuthorize("hasAuthority('ADMIN')")
 	public String adminProfile() {
 		return "Welcome to Admin Profile";
 	}
 
-	@PostMapping("/generateToken")
+	@PostMapping("/login")
 	public String authenticateAndGetToken(@RequestBody AuthRequest authRequest) {
 		Authentication authentication = authenticationManager.authenticate(
-				new UsernamePasswordAuthenticationToken(authRequest.getUsername(), authRequest.getPassword()));
+				new UsernamePasswordAuthenticationToken(authRequest.getEmail(), authRequest.getPassword()));
 		if (authentication.isAuthenticated()) {
-			return jwtService.generateToken(authRequest.getUsername());
+			return jwtService.generateToken(authRequest.getEmail());
 		} else {
 			throw new UsernameNotFoundException("Invalid user request!");
 		}
