--- backend/src/main/java/app/backend/dto/UserInfoDTO.java
@@ -1,59 +1,0 @@
-package app.backend.dto;
-
-import java.util.UUID;
-
-import app.backend.enums.RoleEnum;
-
-public class UserInfoDTO {
-
-	private UUID id;
-
-	private String name;
-
-	private String email;
-
-	private String password;
-
-	private RoleEnum role;
-
-	public UUID getId() {
-		return id;
-	}
-
-	public void setId(UUID id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-	public RoleEnum getRole() {
-		return role;
-	}
-
-	public void setRole(RoleEnum role) {
-		this.role = role;
-	}
-
-}
