+++ backend/src/main/java/app/backend/entity/AuthRequest.java
@@ -1,27 +1,35 @@
 package app.backend.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
 public class AuthRequest {
 
-    public String getUsername() {
-		return username;
+	private String email;
+	private String password;
+
+	public AuthRequest() {
+		super();
 	}
+
+	public AuthRequest(String email, String password) {
+		super();
+
+		this.email = email;
+		this.password = password;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
 	public void setEmail(String email) {
 		this.email = email;
 	}
+
 	public String getPassword() {
 		return password;
 	}
+
 	public void setPassword(String password) {
 		this.password = password;
 	}
-	private String username;
-    private String password;
 
 }
